Using a more global ip address instead of 127.0.0.1 or ::1.
When the attester server is bound to all network interfaces
(with the 0.0.0.0 or :: address), attester previously sent
the internal 127.0.0.1 or ::1 address to browser launchers.
If launchers send this address to remote browsers it does not work.
Now, attester uses the first non-internal ip address returned
by os.networkInterfaces() which will probably work better with
remote browsers in most cases.