Stores which expected browsers a connected slave matches
This change allows slave controllers to check whether a connected slave
matches an expected browser. Also, the process to assign tasks is a bit
simpler: it is no longer needed to check at task assignment time whether
a slave matches a browser from the campaign, as the matching browsers are
already available.